:recycle: Refactor: 디렉토리 및 파일 이름 변경
