+++ src/main/java/com/umc/TheGoods/service/OrderService/OrderCommandService.java
@@ -1,9 +1,9 @@
 package com.umc.TheGoods.service.OrderService;
 
 import com.umc.TheGoods.domain.order.Orders;
-import com.umc.TheGoods.web.dto.order.OrderRequest;
+import com.umc.TheGoods.web.dto.order.OrderRequestDTO;
 
 public interface OrderCommandService {
 
-    public Orders create(OrderRequest.OrderAddDto request, Long memberId);
+    public Orders create(OrderRequestDTO.OrderAddDto request, Long memberId);
 }
