+++ src/main/java/com/umc/TheGoods/web/dto/member/MemberResponseDTO.java
@@ -1,4 +1,4 @@
-package com.umc.TheGoods.web.dto.Member;
+package com.umc.TheGoods.web.dto.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
