+++ src/main/java/com/umc/TheGoods/web/dto/order/OrderResponseDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class OrderResponse {
+public class OrderResponseDTO {
 
     @Builder
     @Getter
